+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 * Pages based on object ids are now formed correctly
 
+### Changed
+* All errors returned to the requestor are standardized
+
 ## [1.1.1] - 2015-08-10
 ### Fixed
 * Build dist for changes